requirements.txt
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
decorator==4.4.2
fsspec==0.7.4
gcsfs==0.6.2
google-api-core==1.20.0
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
idna==2.9
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
protobuf==3.12.2
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.15.0
urllib3==1.25.9
xlrd==1.2.0
